Change the base class of Scene to OrderedUpdates so now it is more easy to set the order of the sprites.
Add comments to Sprite class.
Change the __init__ signature to easy the initialization with a given
texture, but keep the old signature working as well.
Add the update_image method to update the internal image while updates
the sprite's size too.
Change the load_image to use the update_image method.
Change the move method to use an underlying method from Rect.